feat: added some documentation
