+++ glyphr/src/font.rs
@@ -1,5 +1,17 @@
+//! # font.rs
+//!
+//! Contains structures used to describe generated fonts
+
 use crate::GlyphrError;
 
+/// Defines how the glyphs are stored in the bitmaps
+#[derive(Clone, Copy)]
+pub enum BitmapFormat {
+    SDF,
+    Bitmap,
+}
+
+/// Contains informations that are bound to the single glyph
 pub struct Glyph<'a> {
     pub character: char,
     pub bitmap: &'a [u8],
@@ -10,12 +22,7 @@ pub struct Glyph<'a> {
     pub advance_width: i32,
 }
 
-#[derive(Clone, Copy)]
-pub enum BitmapFormat {
-    SDF,
-    Bitmap,
-}
-
+/// Contains informations that are useful for every glyph
 #[derive(Clone, Copy)]
 pub struct Font<'a> {
     pub glyphs: &'a [Glyph<'a>],
@@ -26,6 +33,7 @@ pub struct Font<'a> {
 }
 
 impl<'a> Font<'a> {
+    /// Returns a Result, Glyph if it's Ok, Err if the glyph is not found
     pub fn find_glyph(&self, ch: char) -> Result<&Glyph<'a>, GlyphrError> {
         self.glyphs
             .binary_search_by_key(&ch, |g| g.character)
@@ -34,6 +42,7 @@ impl<'a> Font<'a> {
     }
 }
 
+/// Used to describe alignment on X axis
 #[derive(Clone, Copy)]
 pub enum AlignH {
     Left,
@@ -41,6 +50,7 @@ pub enum AlignH {
     Right,
 }
 
+/// Used to describe alignment on Y axis
 #[derive(Clone, Copy)]
 pub enum AlignV {
     Top,
