+++ glyphr/src/lib.rs
@@ -4,13 +4,13 @@
 
 #![no_std]
 
-pub mod font;
-pub mod renderer;
-pub mod sdf;
-pub mod utils;
+mod font;
+mod api;
+mod renderer;
+mod utils;
 
 pub use font::{AlignH, AlignV, BitmapFormat, Font, Glyph};
 pub use glyphr_macros::{generate_font, generate_fonts_from_toml};
-pub use renderer::{
+pub use api::{
     BufferTarget, Glyphr, GlyphrError, RenderConfig, RenderTarget, SdfConfig, TextAlign,
 };
